% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split2.R
\name{split2}
\alias{split2}
\title{Simulating 2 splits}
\usage{
split2(
  data,
  est.lambda = FALSE,
  lambda.val = 0,
  n.exposure,
  n.outcome,
  n.overlap,
  cor.xy,
  pi = 0.5,
  mr_method = "mr_ivw",
  threshold = 5e-08
)
}
\arguments{
\item{data}{A data frame to be inputted by the user containing summary
statistics from the exposure and outcome GWASs. It must have at least five
columns with column names \code{SNP}, \code{beta.exposure},
\code{beta.outcome}, \code{se.exposure}, and \code{se.outcome}. Each row
must correspond to a unique SNP, identified by \code{SNP}.}

\item{est.lambda}{A logical value.}

\item{lambda.val}{A numerical value.}

\item{n.exposure}{A numerical value to be specified by the user which is equal
to the number of individuals that were in the exposure GWAS.}

\item{n.outcome}{A numerical value to be specified by the user which is equal
to the number of individuals that were in the outcome GWAS.}

\item{n.overlap}{A numerical value to be specified by the user which is equal
to the number of individuals that were in both the exposure and outcome GWAS.}

\item{cor.xy}{A numerical value to be specified by the user which is
equal to the observed correlation between the exposure and the outcome. This
value must be between -1 and 1.}

\item{pi}{A numerical value. The default setting is \code{pi=0.5}.}

\item{mr_method}{A string. The default setting is \code{mr_method="mr_ivw"}.}

\item{threshold}{A numerical value. The default setting is
\code{threshold=5e-8}.}
}
\value{
A data frame
}
\description{
Simulating 2 splits
}
