% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_lambda.R
\name{est_lambda}
\alias{est_lambda}
\title{Estimating lambda}
\usage{
est_lambda(data, z.threshold = 0.5)
}
\arguments{
\item{data}{A data frame to be inputted by the user containing summary
statistics from the exposure and outcome GWASs. It must have at least five
columns with column names \code{SNP}, \code{beta.exposure},
\code{beta.outcome}, \code{se.exposure}, and \code{se.outcome}. Each row
must correspond to a unique SNP, identified by \code{SNP}.}

\item{z.threshold}{A value}
}
\value{
A value
}
\description{
Estimating lambda
}
