% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_simss.R
\name{mr_simss}
\alias{mr_simss}
\title{Main function for MR-SimSS}
\usage{
mr_simss(
  data,
  n.exposure,
  n.outcome,
  n.overlap,
  correlation,
  n.iter = 100,
  splits = 2,
  pi = 0.5,
  pi2 = 0.5,
  threshold = 5e-08,
  mr_method = "mr_ivw",
  parallel = TRUE,
  n.cores = NULL,
  fun1 = split2,
  fun2 = split3
)
}
\arguments{
\item{data}{A data frame to be inputted by the user containing summary
statistics from the exposure and outcome GWASs. It must have at least five
columns with column names \code{SNP}, \code{beta.exposure},
\code{beta.outcome}, \code{se.exposure}, \code{se.outcome} and
\code{eaf.exposure}. Each row must correspond to a unique SNP, identified by
\code{SNP}.}

\item{n.exposure}{A numerical value to be specified by the user which is equal
to the number of individuals that were in the exposure GWAS.}

\item{n.outcome}{A numerical value to be specified by the user which is equal
to the number of individuals that were in the outcome GWAS.}

\item{n.overlap}{A numerical value to be specified by the user which is equal
to the number of individuals that were in both the exposure and outcome GWAS.}

\item{correlation}{A numerical value to be specified by the user which is
equal to the observed correlation between the exposure and the outcome. This
value must be between -1 and 1.}

\item{n.iter}{A numerical value which specifies the number of iterations of
the method, i.e. the number of times sample splits are randomly simulated.
The default setting is \code{n.iter=1000}.}

\item{splits}{A numerical value that must be equal to 2 or 3, indicating
whether splits of 2 or 3 should be simulated. It is recommended that in the
case of no overlap between the two GWASs that splits of 2 should be used
while in the presence of overlap, especially full overlap, splits of 3 should
be used. The default setting is \code{splits=2}.}

\item{pi}{A numerical value . The default setting is \code{pi=0.5}.}

\item{pi2}{A numerical value . The default setting is \code{pi2=0.5}.}

\item{threshold}{A numerical value. The default setting is
\code{threshold=5e-8}.}

\item{mr_method}{A string. The default setting is \code{mr_method="mr_ivw"}.}

\item{parallel}{A logical value . The default setting is \code{parallel=TRUE}.}

\item{n.cores}{A numerical value . The default setting is \code{n.cores=NULL}.}

\item{fun1}{The function used in the 2-split approach. This should not be
changed from \code{fun1=split2}.}

\item{fun2}{The function used in the 3-split approach. This should not be
changed from \code{fun2=split3}.}
}
\value{
A list
}
\description{
\code{mr_simss} is a function which is based on simulated sample splitting in
order to alleviate Winner's Curse bias in MR causal effect estimates. It also
takes into account sample overlap between the exposure and outcome GWASs. It
uses GWAS summary statistics together with knowledge of the number of
overlapping samples and the correlation between the exposure and the outcome.
It works in combination with existing MR methods, such as IVW and MR-RAPS.
}
