% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split2.R
\name{split2}
\alias{split2}
\title{Simulating 2 splits with MR-SimSS}
\usage{
split2(data, lambda.val = 0, pi = 0.5, mr_method = "mr_ivw", threshold = 5e-08)
}
\arguments{
\item{data}{A data frame to be inputted by the user containing summary
statistics from the exposure and outcome GWASs. It must have at least five
columns with column names \code{SNP}, \code{beta.exposure},
\code{beta.outcome}, \code{se.exposure} and \code{se.outcome}. Each row must
correspond to a unique SNP, identified by \code{SNP}.}

\item{lambda.val}{A numerical value which is computed within the main function, \code{mr_simss}.
It is an estimate of \emph{lambda}, a term used to describe the correlation between the SNP-outcome and
SNP-exposure effect sizes. The default setting is \code{lambda.val=0}.}

\item{pi}{A numerical value which determines the fraction of the first split. This is the fraction that will be used
for SNP selection. The default setting is \code{pi=0.5}.}

\item{mr_method}{A string which specifies the MR method that MR-SimSS works in
combination with. It is possible to use any method outputted in the list
\code{TwoSampleMR::mr_method_list()$obj}. However, it is currently advised
that the user chooses \code{"mr_ivw"} or \code{"mr_raps"}. The default
setting is \code{mr_method="mr_ivw"}.}

\item{threshold}{A numerical value which specifies the threshold used to
select instrument SNPs for MR at each iteration. The default setting is
\code{threshold=5e-8}.}
}
\value{
A data frame which contains
the output from one iteration. It is in a similar style as the output from
using the function \code{mr} from the \code{TwoSampleMR} R package. The MR method used, the number of instrument SNPs, the causal effect estimate, it associated standard error and \emph{p}-value are all outputted.
}
\description{
\code{split2} is a function which is used by the main function, \code{mr_simss} in order to perform the 2 split approach of the method,  \strong{MR-SimSS}.
}
